+++ src/main/java/com/verdantartifice/primalmagic/common/entities/projectiles/SpellProjectileEntity.java
@@ -14,7 +14,7 @@ import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.projectile.ThrowableProjectile;
+import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.EntityHitResult;
@@ -27,13 +27,13 @@ import net.minecraftforge.fmllegacy.network.NetworkHooks;
  * 
  * @author Daedalus4096
  */
-public class SpellProjectileEntity extends ThrowableProjectile {
+public class SpellProjectileEntity extends AbstractArrow {
     protected static final EntityDataAccessor<Integer> COLOR = SynchedEntityData.defineId(SpellProjectileEntity.class, EntityDataSerializers.INT);
     
     protected final SpellPackage spell;
     protected final ItemStack spellSource;
     
-    public SpellProjectileEntity(EntityType<? extends ThrowableProjectile> type, Level worldIn) {
+    public SpellProjectileEntity(EntityType<? extends AbstractArrow> type, Level worldIn) {
         super(type, worldIn);
         this.spell = null;
         this.spellSource = null;
@@ -102,11 +102,22 @@ public class SpellProjectileEntity extends ThrowableProjectile {
 
     @Override
     protected void defineSynchedData() {
+        super.defineSynchedData();
         this.getEntityData().define(COLOR, 0xFFFFFF);
     }
 
     @Override
     public Packet<?> getAddEntityPacket() {
         return NetworkHooks.getEntitySpawningPacket(this);
+    }
+
+    @Override
+    protected float getWaterInertia() {
+        return 0.99F;
+    }
+
+    @Override
+    protected ItemStack getPickupItem() {
+        return ItemStack.EMPTY;
     }
 }